artifacts/yolo-service/requirements.txt
opencv-python-headless
fastapi
uvicorn[standard]
numpy
psycopg2-binary
